requirements.txt
# API
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Base de données
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# MQTT
paho-mqtt==1.6.1

# IA
numpy==1.24.3
pandas==2.1.4
scikit-learn==1.3.2
joblib==1.3.2
mindspore

# Utils
python-multipart==0.0.6
requests==2.31.0